[pulse][4/??] revise models of std::basic_string
Summary:
This is needed to remove the weirdness that is string evaluation: evaluating a
string constant would store the string in a field in memory instead of treating
strings as pure values. I don't think we need to do that but I think C++ models
rely on it so let's migrate them.

This is an attempt to make the C++ represent basic strings as having an int
field and a pointer field to a pure string value.

Reviewed By: skcho

Differential Revision: D54317282

fbshipit-source-id: 42017ff88b697938b2dba7ec7e446dfa1ea86793