[pulse][6/??] use constants we already know when normalising terms
Summary:
If we knew "x = y^z" and *then* learned y="foo" and z="bar" we would correctly
normalise x="foobar" but not the other, simpler way around. Fix it by looking
up constant equalities when normalising terms for the first time, just as we do
for numerical constants already.

A C++ test becomes FP (again, it was FP at the base of the stack too).
This is due to a subtle treatment of unknown functions and quantifier
elimination:

```
std::optional<std::string> getOptionalValue() {
  auto value = std::getenv(envVar);
  if (value) {
    return std::string{value};
  }
  return std::nullopt;
}
```

Here `getEnv` is unknown and `envVar` is a global with value
`"ENV_VAR"`. We thus get `value = getEnv("ENV_VAR")` but because that's
a local variable we end up dropping all of this on the floor during
summary creation. Previously we got `value = getEnv(envVar)` and since
`envVar` is live so was the value of `value`, hence we kept `value=1`
(for the true branch, say) in the conditions. This is important because
the FP is due to taking a branch `getenv(envVar) == 0`, which should not
be feasible.

This seems tricky to fix! We got rid of the FP sort of by chance so
thinking of a fix is left as future work.

Reviewed By: skcho

Differential Revision:
D54903864

Privacy Context Container: L1208441

fbshipit-source-id: f58d8d1519c5f6b29b0c0d7945be74397e44f795