[pulse][7/??] ground terms should not make issues latent
Summary:
We should never get ground terms in conditions (they should be
evaluated to true/false instead) but if we do we should ignore them.

At the moment there is a problem with Java models for strings: string
constants do not get converted to objects as they should and so the
models get confused. For example:

```
String s = "abc"; // should get s as an object String with fields for
                  // content and length, but instead we get s = "abc"
if (TextUtils.isEmpty(s)) ...
  // we get "abc" = 0 in the precondition for the s == null case
```

We could also decide to raise Unsat in those cases instead, it's unclear
which is best at the moment.

The real fix would be to fix the translation so that the models work as
intended.

Reviewed By: davidpichardie

Differential Revision: D54953682

fbshipit-source-id: 0c1e1ab829545ffb80a27856e8c918e3d31565dd