[report] Add categories to issue types
Summary: We create a category field for issue types that can be exported in the report.json. The categories are listed, and the issue types will be filled in following diffs. Because there are too many issue types that we are deprecated or not in active development or production, we make the field optional and add it only to relevant issue types.

Reviewed By: skcho

Differential Revision: D55188067

fbshipit-source-id: 0766c355b7630c27493d8f0bf8a261c480ba1fbb