[refactor] Lineage: Merge update_write and add_edge
Summary:
Small Lineage refactor. The Lineage domain has two functions before this diff:

- `add_edge` adds an edge to the (partial) graph. To be used when any flow must be recorded.
- `update_write` is used to update the nodes where a cell was last written, to be used additionnally
   to `add_edge` when recording flow to a cell (ie. a variable field).

We merge them into `add_flow_edge` that will add the flow and, if the destination is a cell, update
`last_writes` accordingly. The rationale is to make the functions that record some flow more
uniform (so they can be refactored later on).

Reviewed By: rgrig

Differential Revision: D55208330

fbshipit-source-id: 18d41fbf5c1ef074b67fa410718724fcf6e9745d