[dead store] Add dummy loads to avoid false positives
Summary:
The liveness checker reported false positives on the arguments of
`__atomic_compare_exchange`. The arguments are needed to call the
library function, even if developers do not need to read them after
the call.

This diff adds dummy loads to avoid dead store false positives on the
library function calls.

Reviewed By: dulmarod

Differential Revision: D55326896

fbshipit-source-id: ac952bc9ed789ba80297a7e850858fa02c6df7f9