Lineage: represent edge sources as sequences of vertices
Summary:
The pre-diff API for adding flow edges to Lineage relies on sources as an (abstract) alias of
vertices, and define a variety of flow-recording functions that, according to the sources, will add
the edges from all the corresponding vertices (eg. all the known subfields of a variable, or multiple
free locals from an expression).

Similarly, destination of edges are built as vertices, and there is another set of specific functions
to record flows into a variable (which is currently the only occurrence of flows with "several
destination vertices").

This does not scale well with the planned introduction of other multiple-vertices sources and
destinations, such as argument nodes of callee procedures (one will have to write multiple new
functions for each new source).

We start refactoring this by offloading the responsibility of traversing several nodes into the
source type itself. Sources are now built as a set of vertices, and we only need one `add_flow`
function that will use the source to build the individual edges accordingly.

We don't touch the destination type yet to make this diff shorter, therefore we still have
a (smaller) variety of `add_write` functions to write into variables.

Reviewed By: rgrig

Differential Revision: D55321253

fbshipit-source-id: 0348eb9b198c26b72eefaa6f8d7695faa5bd0b80