Lineage: represent destinations as sequences of nodes
Summary:
We go further in the refactoring of flow by also representing destination as vertices sequences.

This allows standardising a bit the flow functions by making some `add_write` redundant. We still
keep a specific set of functions writing from/to variable paths that make use of the information
carried by the underlying cells. The plan is to also make this pattern generic to be able to extend
it to other nodes than variable paths.

Reviewed By: rgrig

Differential Revision: D55433777

fbshipit-source-id: 150e5679cd5c7a104e6faa8b70c1bac012438370