[transitive-acces] take more dropped disjuncts into account
Summary:
The transitive-access analysis keep track of transitive callees in Pulse disjunctive states
but also in the non-disjunctive state. The later is used whenever a disjunct is dropped to not completely
lose the corresponding transitive callees (with a risk of not detecting infeasible call graph paths).

We were not taking into account all dropped disjuncts. This diff fixes this.

Reviewed By: skcho

Differential Revision:
D55189508

Privacy Context Container: L1208441

fbshipit-source-id: c77dedcf88318c60c18faa30122f1344b007e8bd