[pulse][RFC] count reachable nodes with 0 disjuncts
Summary:
we only count nodes reachable from the start node,
not following exceptional edges.

Hence we use a simple depth-first-search.

It we count nodes in catch blocks, this makes difficult to understand the results because this is expected that our analysis will never enter in some of them.

Reviewed By: nbenton

Differential Revision:
D55082611

Privacy Context Container: L1208441

fbshipit-source-id: 34fa381bee09550fbd018a81b537b42955329f5a