[pulse] Revise make_shared semantic model
Summary:
In `std::make_shared` semantic model, it tries to find a proper constructor depending on argument
types.  But, for a lambda type argument, i.e. `std::function`, the type was changed to an artificial
class `lambda_*`, thus could not find the constructor.  This diff revises the type comparison of the
lambda, so that `lambda_*` type is matched to `std::function`.

Reviewed By: dulmarod

Differential Revision: D55797333

fbshipit-source-id: e2404cf82b83f340c88307df13f73c2957a53ae2