[clang] Implement the ObjC setters correctly when the property has the copy attribute
Summary:
We implement getters and setters of properties internally, but we had not taken into account the copy attribute. When that is present the setter implementation includes a call to copy

```
setField(Field* field) {
  self->_field = [field copy];
}
```

This is used in blocks sometimes, and not implementing it was causing false positive retain cycles.

Reviewed By: skcho

Differential Revision: D55969745

fbshipit-source-id: b69a39bc292c7861dd364bfcab5a40ad0fd25a3a