[clang] Add a test that shows that the copy property and the preanalysis to swap copy with copyWithzone work well together
Summary:
The setter for properties that have the copy attribute includes a call to `NSobject.copy`, but there is another preanalysis that swaps this call to the correct call to `copyWithZone:` method of the given class. This is because that's what `NSObject.copy` does internally.

So here we are adding a test that shows that the swap works correctly.

Reviewed By: skcho

Differential Revision: D56012605

fbshipit-source-id: 3c9cea76c19294421fbee66eb0e047ae4163b9f2