make Hack access comparison ignore the classname part
Summary: We observed Hack field accesses to the same field occurring sometimes with a wildcard classname and sometimes with a proper classname. The result was the creation of multiple fields with the same (string) name, instead of an update. Since there's no particular reason to tag Hack fieldnames with a class, and it's actually hard for hackc to always assign the "correct" classname (type aliases, for example), we instead make `Fieldname.compare_name` (which is used in `Access.compare`) ignore the classname for Hack fields.

Reviewed By: skcho

Differential Revision: D56304036

fbshipit-source-id: 53537734df05c63a200b086ebc905eaaf2839f6a