[dict missing key] Report issue only on dict parameter
Summary:
Currently, the frontend cannot distinguish dict and shape, which is introducing many false positives
on shape, esp. with aliases.  To avoid that, this diff reports issues only when the the object is a
parameter and its type is `HackDict`.

This is a temporary heuristic and we will work on the frontend to translate more information about
alias.

Reviewed By: geralt-encore

Differential Revision: D56639780

fbshipit-source-id: 4ed717fed7034eea2bbac212db5ccdd87e23bd13