[cxx_ref] Assume that the block is no escaping if it is not passed to a function
Summary:
If a block is not passed to a function, but executed in the same scope, then we should assume it is no escaping, and thus, we should not report C++ Ref Captured In Block issue, because that isn't a problem in that case.

It is also possible that the block is stored in an instance variable instead, but we are assuming for now that this happens always via a function call that is annotated with NO_ESCAPE if needed, since that seems to be the most common case.

Reviewed By: geralt-encore

Differential Revision: D56702415

fbshipit-source-id: 3e010f83cfcff9274a8ee372281fef4e0c79ac02