[retain cycles] Improve the filtering of retain cycles reporting
Summary:
Before we were a bit inconsistent when filtering the objects that can make the cycle that we report: we were checking only some of the objects in the cycle, now we check the properties in all the objects.

The properties are:
1. Non nil
2. Not part of a cycle previously reported
3. Known (We have a decompiler value for it)
4. We only take into account ObjC classes or blocks

This means that we can be sure that we won't report on cycles that include structs or C++ classes.

Reviewed By: skcho

Differential Revision: D56636988

fbshipit-source-id: 6c5bb8ce17c10b1f79b905139917e4e76b0593ff