[refactor] Pull `Lineage.G.Out` from `Lineage.G` into a top level `Lineage.Out` module
Summary:
`Lineage.G.Out` is the module managing Lineage json exports. It has some subtle stateful invariants
(in particular on channel creation and uniqueness) that make it non-trivial to extend it in an
obviously-correct way to export pathfinding results.

This diff is a first step in a small refactoring series that aims at making the aforementioned change
easier to review.

Reviewed By: rgrig

Differential Revision: D56800224

fbshipit-source-id: 232a1ec55b20553462c3790a90343f84e675bedc