[refactor] Lineage.Out: have base json writing functions take channel as an explicit argument
Summary:
As a next step to untangle the stateful aspects of Lineage Json export, we have the core json-writing
functions take the out channel as an explicit argument rather than calling the global `channel ()`
function.

This global function (that returns a unique outchan per pid, closed at the end of the process) is now
only called on the final `Summary.report` function.

Reviewed By: rgrig

Differential Revision: D56800223

fbshipit-source-id: a5786f99d45e5b9d251b801d6fcd899623e07fa2