[retain cycles] Adding a test that shows an example where we report a wrong cycle
Summary:
The issue is a FP because `_component` has the `__weak` attribute. I was able to identify the problem: it's because we somehow skip the function `NSObject.new`, that means that after line 41 we wrongly identify the objects `c` and `bridgingDelegate`, which leads to wrong reports.

Seems to be a wrong translation of `[super new]`. Will send a fix soon.

Reviewed By: geralt-encore

Differential Revision: D56877643

fbshipit-source-id: bb1976049574cefdbf395cfcf2ed75844f38327f