[annot] Add option to skip annotations from superclasses/interfaces
Summary: Previously, if a class inherited from a base class or extended an interface, all of their annotations applied to all of the methods of the class. With this option, only "local" annotations are applied. That is, a function `foo` is only considered to be annotated by the interface or base class annotation, if `foo` is present in the base class / interface.

Reviewed By: ngorogiannis

Differential Revision: D56885801

fbshipit-source-id: 339858fa288018cbdd62a950057556e9784c6925