[clang] Improve broken translation of [super new]
Summary: For `[self new]` we translate it as `[[self alloc] init]`, but somehow missed to do the same with `[super new]`. This was causing `NSObject.new` to be skipped and because of this Pulse was considering unrelated objects to be equal. That was the reason for the false positive retain cycle in the tests that has now been fixed.

Reviewed By: skcho

Differential Revision: D56882442

fbshipit-source-id: f582971ae1d9b2f95cc4aba691be44ff31ebeb73