[refactor] Lineage: have Out depend on Summary instead of the opposite
Summary:
Have the `Out` exporting module depend on Summary makes it easier to add other exporting
functions (such as exporting only graphs, adding output formats and what not).

As it is large in lines of code it is easier to read as a single commit. This moves code around and
changes use of `Lineage.Summary.report` into `Lineage.Out.report_summary`.

Reviewed By: rgrig

Differential Revision: D56899266

fbshipit-source-id: 4164b02cd80573fb98510725a13e927f6675acaa