Lineage taint pathfinding prototype
Summary:
This diff implements a prototypal version of Lineage pathfinding, working on one source and sink.

The corresponding workflow is to run a Lineage analysis first, then use `infer
report --lineage-source [...] --lineage-sink [...]`. The lineage results will be reused and several
pathfinding requests can therefore be run subsequently without rerunning the lineage analysis (as
long as the code itself isn't modified).

It works by first collecting a subgraph of the Lineage graph composed of all reachable nodes from the
source, then from this subgraph will collect the nodes from which the sink can be reached.

It has limitations that will likely prevent its uses on real-life production code (for instance, it
will not find flows involving procedures with no lineage model, such as some native stdlib ones; it
also doesn't support field-sensitive runs of the Lineage analysis), but is sufficient to run
experiments in a controlled environment.

Reviewed By: rgrig

Differential Revision: D56899263

fbshipit-source-id: 616573c2641fed4d5e2d0e640bb0fb31e790e15f