Make LineageTaint more robust to unexpected usecases
Summary:
The prototype of LineageTaint has a lot of cases where it raises an exception on legit usecases that
happen to derive slightly from the ideal one.

We fix most of them: pathfinding from/to functions without summary, do not raise when the final graph
is empty, procedures whose presence in the flow depends on the context.

Reviewed By: rgrig

Differential Revision: D57057458

fbshipit-source-id: edcf218f35bf6c9fc9e24c5c0790a4abeae420d1