[infer] add nullable field to sizeof_data
Summary:
First step towards clarifying the relationship between Infer's dynamic types and nullability (particularly revelant to Hack, which can say either `is C` or `is ?C`). We add a flag to the (poorly named) sizeof_data type so that when we translate an instanceof operation we can indicate if the type being tested against should include null or not.

(The concrete booleans that are added, but not yet used, in this diff may require changing later.)

Reviewed By: geralt-encore

Differential Revision: D56883390

fbshipit-source-id: f77cb8ac4e6ae5bdb331f1e34bd6e8be082b2155