[annot] Warn when a method is both a source and a sink
Summary: Add support for reporting the corner case where some method is marked both as a source and a sink. Previously this was not reported because in this case there is no `call` instruction happening. This can be enabled by a flag which is `false` by default.

Reviewed By: rgrig

Differential Revision:
D57156089

Privacy Context Container: L1208441

fbshipit-source-id: e81f6248d1d70fbaa22000d810f693594cdfee83