[report] Adding a flag to specify false positives to skip
Summary:
Sometimes users know that Infer found a false positive and would like to suppress it. Now we have a way. We can't use line to specify the issue because that changes too quickly, we we specify bug_type, file, class and method.

We could also add some relevant variable names or such in the future.

Reviewed By: jvillard

Differential Revision: D57112894

fbshipit-source-id: ab69a5d121e70acc56f0982485bb3956bd4b7eec