[pulse] Bug-fix nontermination
Summary:
Suppose a formula contained `(x<2)=a1 && a1=0`, and we want to prune
`x=2`. Pulse used to notice that the inequality is the same as `x≥2`,
which it tried to represent as `x-2=a2`, where a2 is a fresh nonnegative
variable. Because `x=2` is known, the `x-2` had normal form `0`, so
Pulse ends up setting `a2=a1=0`. While doing so, it notices again that
`(x<2)=a1` has a simpler form, and starts again.

The fix in this commit is to encode `l1≥l2` as `l1-l2=a` (for fresh
nonnegative `a`) iff the normal form of `l1-l2` is not already known to be
nonnegative (or known to be negative, in which case a contradiction is
detected).

Reviewed By: jvillard

Differential Revision: D57113153

fbshipit-source-id: a90d095fedb94c665224587e384951c623a3616e