Lineage: introduce slots for sources and destinations
Summary:
We enrich Lineage sources and destinations used to record flow from sequences of vertices to
respectively sequences of vertex and slot and sequences of vertices given a slot.

Rather than having N² flow recording functions depending on the particularities of theses sources and
destinations, this is a foundation for a unified way of connecting sources and destinations
together. For instance, "parallel" flow will correspond to use the slot generated with each source
vertex to produce the corresponding destination ones, and "product" variable path flow amounts to
ignoring the slots of the source vertex (which is done on the source itself using `Src.oblivious`)
and use the root field path `[]` as a slot for destination (which is done on the destination using
`Dst.all_fields`).

`Argument` and `Return` source and destinations can now be considered as having a field path
slot. Since the flow production is source-directed, we add to `LineageShape` a function to fold over
the terminal fields of an argument, but this isn't necessary yet for the return (as we use a normal
variable to produce the slots).

We do not transform every flow-production code yet into this new model: to ease reviewing and
checking for specific corner cases, some of the old functions (eg. write from/to var paths) remain.

Reviewed By: rgrig

Differential Revision: D57212632

fbshipit-source-id: dd4b7356898026cfddd3166b2c526373196e916c