Lineage: remove some use of add_parallel_var_path_flow
Summary:
`add_parallel_var_path_flow` used to be an hard-coded double-cell-fold on two variable paths to write
in parallel, also allowing the user to provide an `exclude` function to not write on some specific
source/destination fields. That exclude function is currently used only on maps:put-like models to
not write flow from the old map on the key being put.

We change all the uses of this function to use the simpler `add_var_path_flow` that is a simple
adapter of the slot-using flow writing function implemented on previous commits (and adds a check for
shape equality for bug catching).

Parallel write with exclusion is kept as-is to make reviewing easier.

Reviewed By: rgrig

Differential Revision: D57217127

fbshipit-source-id: 2c7d0fa6ec93b713f807d1d9b1833e00fe5b4241