Lineage: have destinations also generate signals
Summary:
Lineage sources generate signals that are used as slots to build the corresponding destinations. That
works well for most flow recording but is not very convenient when the destination is holding some
information, which is the usecase for the ad-how `add_flow_to_var_path`.

We add signals to destinations, which allows to implement a single generic `add_flow_f` function that
can use the signals from both the source and the destination to generate the kind. It subsumes the
previous functions to write from and to variable paths into special nodes.

Reviewed By: rgrig

Differential Revision: D57280329

fbshipit-source-id: 88015351ea23a3f00e68b4da7e16d7d74fe4ef7a