[infer] Add PotentialInvalidSpecializedCall to pulse access result
Summary: This is a preparation step to make sure that we can report "Cannot instantiate abstract class" issues at correct place. The problem is that initially we were reporting the issue as soon as we discover it which is technically correct but not very useful. The main idea is to delay issue reporting until we see where specialization has happened and report it there. By doing that we'll be able to build correct calling context and report the issue where problematic call is actually appearing.

Reviewed By: jvillard

Differential Revision:
D57381711

Privacy Context Container: L1122176

fbshipit-source-id: b414742abe5a3d9c0da7e52daf644211c824fe50