Lineage: add field exclusion as Src and Dst functions, factoring add_filtered_var_path_flow out
Summary:
`add_filtered_var_path_flow` was the last remnant of ad-hoc source/dest-specific functions after the
last refactoring passes. It is specifically used for the `maps:put` model to avoid writing flow from
the put-key element of the source map.

To factor it away from the code, we add field-exclusion functions within the `Src` and `Dst`
module. It then appears as another instance of the general flow recording function. To keep checking
for shape equality, we implement this by adding to `add_var_path_flow` a new `exclude` optional field
label argument.

Reviewed By: rgrig

Differential Revision: D57280338

fbshipit-source-id: 17e573b5d509b28f0da4ec4cac1c66e99ed6218b