Lineage: split ArgumentOf/ReturnOf nodes on their known fields
Summary:
Within a procedure analysis, the Argument and Return nodes hold some field information. The
ArgumentOf and ReturnOf nodes, that make the connection to other procedures graphs, used to not be
field-sensitive in the caller graph.

This diff makes ArgumentOf and ReturnOf hold shape informations. This allows in turn to have all the
argument and return nodes splitted in the json output, making the graph simpler to traverse (because
it is less connected).

Injection and projection metadata now only exist on some Call and Return edges. This happens on
shape-polymorphic callees when they're called with shapes having fields.

Reviewed By: rgrig

Differential Revision: D57446503

fbshipit-source-id: 0e50279bfcbe0149c95d27a463022d78a9ca2b9a