[retain cycles] Add a new issue type to keep track of potential retain cycle where we are missing data about weak access
Summary:
This is for debugging purposes mostly, we want to be able to see how often this is happening and debug those cases with the aim of finding a solution.

The reason why we are missing the data is the fact that we have multiple tenvs with different types in each. So we may be missing the required types in the tenv at hand when we are checking for the cycle.

Reviewed By: jvillard

Differential Revision: D57570974

fbshipit-source-id: 7f28ad10319a44bb3338a421ff593359d4331f42