Revert "[pulse][hack] slightly simpler and more complete hhbc_cmp_eq model when the values are equal"
Summary:
This reverts commit "[pulse][hack] slightly simpler and more complete
hhbc_cmp_eq model when the values are equal".

Problem: `prune_eq x y` may be misleading as assigning a value to `x`
when it is given as a parameter.

As an example, when a parameter `x` is pruned by `prune_eq x y`,

```
void foo($x) {
// post: $x -> -
// pre: $x -> -
n$23 = "abc";
// n$23->val -> "abc"
if (hhbc_cmp_same($x, n$23)) {
  // post: $x->val -> "abc"
  // pre: $x -> -
  ...
}
```

Thus, when calling foo in a caller,

```
$y = "def";
// $y->val -> "def"
foo($y);
// $y->val -> "abc"
```

Reviewed By: jvillard

Differential Revision: D57723624

fbshipit-source-id: 7617443a02cd534621d4a0d638cb37c8bdb003ae