[ondemand] refine the return type of ondemand analysis
Summary:
This allows analyses to know what happened when ondemand comes back with no
summary. Not used for now, except  to make some debug messages better, which is
already a little bit useful for debugging when we get no summary. The plan is
to use this to detect mutually recursive procedures, and perhaps infinite
recursion.

Reviewed By: skcho

Differential Revision: D57723301

fbshipit-source-id: 109dfa5119e2912d771042e0db2165cfc1da4089