[clang] Improve the name of fields given to ObjC properties
Summary:
In the tenv that includes the header of a class, we have info about the properties including attributes but not the precise name of the field that corresponds to the property. That is often decided in the implementation of the class. However, very often the field name is the name of the property with an underscore in front. So we add this as a heuristic.

For checking retain cycles, if we can't find the field, we try with or without the underscore, and proceed.

This solves the issues in the tests.

There are still other missing field issues when the field is defined in the implementation class. For those we do need to try to find the tenv of the implementation.

Reviewed By: ngorogiannis

Differential Revision: D57727010

fbshipit-source-id: a5eed44b982841989a909060c399c2ce43bda23a