[stats] count call sites with infeasible callee sumary
Summary:
we increment this new counter each time a caller
is not able to get any ContinueProgram from a callee.

We have to use globals in order to detect that. We use the `AnalysisGlobalState` module that will make sure we don't mess up global references when the current analysis is interrupted by the ondemand interprocedural system, and restored later.

Reviewed By: skcho

Differential Revision:
D57207070

Privacy Context Container: L1208441

fbshipit-source-id: 9cdf8c5dcbb41049bd4ffdd4cfe2060685099971