[erl-frontend] Add support for maybe expression
Summary:
The [`maybe`](https://www.erlang.org/doc/system/expressions#maybe) expression of Erlang was introduced in OTP 25 and will be enabled by default from OTP 27. So *maybe* it's time to add support for it.

Note that this diff only adds support for the `maybe` expression and not (yet) the `else` clauses and the `?=` operator that go with it. We plan to add support for that in follow-up diffs.

Reviewed By: rgrig

Differential Revision: D57909676

fbshipit-source-id: b9aca86411d30b054c42c1f5134795cc671c9f1e