[pulse] detect mutual recursion cycles of arbitrary length
Summary:
Make the analysis interprocedural. This raises an important issue: where to
report the cycle? Because the procedure where we'll detect the cycle depends on
the analysis schedule, it is non-deterministic. Also, we don't want to miss the
procedure where the cycle is first introduced (by a diff) by reporting it
elsewhere in the cycle. So, let's report the cycle in each place in the cycle.

Reviewed By: skcho

Differential Revision: D57785923

fbshipit-source-id: db8945938badd523c0facb8c8b442af991f27604