[erl-frontend] Add support for map comprehensions
Summary:
We only parsed these before, now we have translation and analysis support as well. Note that translation is precise, but there are two abstractions happening in analysis:
1. Similar to list comprehension, generators are translated via loops, affected by Infer's loop unrolling bound.
2. Maps have a recency abstraction (only storing one -- the most recent -- key/value).

Translation of map comprehensions share much of their code with list comprehensions: generators and filters are the same, the difference is initialization (start with empty list/map), the loop body (adding to map/list) and storing the result (lists need to be reversed).

Also note that map generators are not yet supported.

Reviewed By: thizanne

Differential Revision: D58006521

fbshipit-source-id: 889770ee5842adf1552d515194830abcc6055f40