Use local field information in ArgumentOf/ReturnOf instead of interprocedural
Summary:
Trying to build the ArgumentOf/ReturnOf nodes with interprocedural field information fails in the
presence of mutual recursion. For instance, on `f({a, b}) -> g({a, b}). g({a, b}) -> f({a, b})`, if
the analysis starts with `f`, then `g` won't have access to the shape information of `f` (as
a recursion cycle will be detected and no summary provided) and cannot therefore build the correct
`ArgumentOf` nodes matching the `Argument` ones inside `f`.

The ideal way of working this out would be to have a finer dependency graph on the analysis, allowing
to have an empty summary for the Lineage part of `f` but still have access to its shapes when
computing the Lineage graph of `g`. This change is quite invasive and would require careful design
and implementation.

To circumvent this, we instead change the ReturnOf and ArgumentOf nodes to hold the field information
as it is known by the current procedure. The matching with the Return/Argument nodes in the callee
can then be done on a later step, when all have been computed, ie. in the report phase.

We then revert the splitting of the argument/return nodes in the exported json graph, as it may be
generated during the analysis phase an can therefore not access the required information. We still
generate the field metadata as it was done before. Now that the nodes carry the full field
information of the current procedure, we remove it from the edges as it becomes redundant.

Future work may include generating matching nodes when the callee shape information is known, as it
may have performance benefits by avoiding to load that shape information again later on, and could
also simplify the connectivity of the json graph by splitting some nodes again.

Reviewed By: rgrig

Differential Revision: D57742614

fbshipit-source-id: 647f2e30f4a0c82cec1d7c8c781c31f5821e234f