Refactor LineageShape to use shape instead of var indirections on formal/return
Summary:
LineageShape summaries used to remember the variables associated to formals and returned value of
procedures to be able to fold over their final field paths.

We change it to remember the formal and return shapes directly, instead of going through the var
indirection. This prompts for a slight refactoring of cell/final path folding to be able to start
from a root shape instead of a variable.

The aim of this change is to prepare recording shapes of other procedure concrete arguments and
returns (in particular, arguments might not be single variables), to make interprocedural graph knot-tying
possible in the pathfinding.

Reviewed By: rgrig

Differential Revision: D57941689

fbshipit-source-id: 1469f62d2f3e3589705ae81eaf30164f523aa3fb