[inferhack] translate hhi to textual decls
Summary:
Now we're being more precise with runtime type information, we need properties and the type hierarchy for builtin types (declared in `.hhi` files) as well as for those in the `.php` files we capture. Otherwise we run the risk of pruning feasible paths as well as following infeasible ones This diff contains a little script to extract this information and output it in Textual syntax, and an updated `models.sil` file with the output manually pasted in.
We'll address better automation and other refinements mentioned in the comments later.

Reviewed By: jvillard

Differential Revision: D58090095

fbshipit-source-id: 1bca1789d9c96c6068850be9ed79c2f84df2b295