Field-sensitive Lineage pathfinding
Summary:
We make Lineage pathfinding field-sensitive. The main obstacle was following the interprocedural
jointing nodes, ie. Argument/ArgumentOf/Return/ReturnOf, where there is no one-to-one field
correspondence between procedures.

We solve this by adding to Shapes of a procedure the local information of shapes of callees
args/return as it knows it. For instance, a procedure calling `id(X) -> X.` on a tuple will know some
fields for the argument of `id` that `id` itself doesn't know of. By leveraging the possibility of
folding over the final paths that cover a given varpath, we can then make the joint between
ArgumentOf and Argument nodes.

Reviewed By: rgrig

Differential Revision: D58091878

fbshipit-source-id: c86539b2af02e6d8ab55f467590da4bff2526a89