Add an experimental --lineage-limit option to reduce taint graph size
Summary:
We add a --lineage-limit option that can be used to reduce the size of a lineage taint graph when it
becomes too big to be inspected. This option will effectively arbitrarily remove some flows from the
graph: the result will therefore not be over-approximating anymore (nor it will be
under-approximating because the remaining flows might be FP).

The current implementation limits the number of outgoing edges followed from a vertex. It is unsure
what is the best way of limiting sizes but this seems to give actionable results already on
production-size codebases.

Reviewed By: rgrig

Differential Revision: D58091877

fbshipit-source-id: c1846f43e58bdabafe7f98113876a7d1b2ca2f4b