[annot] Cleanup reporting functions
Summary:
Clean up reporting functions: `report_src_snk_path` called `report_call_stack` and it passed in `f_report` as the `report` argument which called `report_annotation_stack`. This was needed becase in Cxx `f_report` was something different. But now with Cxx support eliminated, we can simplify this.

Plus rename some functions to make it clear what they do. Previously we just used the word "report" everywhere, but some of the functions were actually checking if something is a source, or were looking for paths to sinks and then finally reported.

Reviewed By: jvillard

Differential Revision: D58190717

fbshipit-source-id: eeedb6d058ff22eb6e31a302029c1e1142d3162d