Stop LineageTaint pathfinding at the shortest source/sink
Summary:
We stop propagating the taint reachability dfs after a sink has been reached (and similarly, stop the
codfs on sources).

This has effect on programs with multiple sources that flow into one another, which I believe could
happen with fields but is not the main expected gain.

The main rationale is to simplify the analysis of programs with source<->source loops such as:

```
source(X) -> X.
sink(X) -> X.
flow() -> sink(source(whatever)).
loop() -> source(source(idontcare)).
```

Without this diff, `loop` would add a loop from the source to itself in the Lineage graph. This loop
could grow very large and is usually of little utility for investigating the result.

Other taint analyses guard this behaviour under a CLI flag, that could be a possibility too if we
think this behaviour may be desirable in the future.

Reviewed By: rgrig

Differential Revision: D58301141

fbshipit-source-id: 67f1d01a0738556097b12d17303d472c0b92bd58