[hack] fix a test because hh is clever now
Summary:
the Hack typechecker is now able to find this kind of error, if the closure argument is obvious.
But the DYNAMIC_TYPE_MISMATCH checker should still be usefull when the KeyedTraversable interface is used.

Note: the current reporting of this checker is not at the right place. This is a known issue.

Reviewed By: nbenton

Differential Revision: D58367043

fbshipit-source-id: ccd828c1d2f60de9f0a31745ec77cfcbd04c4f8e