Refactor Lineage edge metadata, adding a dummy Derives
Summary:
In preparation of adding Derived procedure metadata in the json output, we refactor the metadata
generation to an easier to expand architecture.

Because some pipeline and scripts may rely on the current format of the injection and projection
metadata, we do not actually generate Derive metadata yet for easier validation; but have it almost
ready so the actual change adding this will be minimal thus easier to trust.

Reviewed By: rgrig

Differential Revision: D58471473

fbshipit-source-id: c857cca4032293bc191c1b0e478bf6835b06a41a