[IR] Add is_weak field to fieldname to propagate this information from the instructions to Pulse heap.
Summary: This avoids looking up the tenv during execution and avoids the problem that the information may be in the wrong tenv. For made up fields for captured variables the is_weak item was already there, so we lift it up to the `Fieldname.t` type. For other fields, we make sure that we create all the objc ones including the `is_weak` field, and that can be ignored for the rest of fields.

Reviewed By: ngorogiannis

Differential Revision: D58465174

fbshipit-source-id: f6d31204f7fc0926c2fbc99cdff5a8ec3d168dc6