[infer] Detach analysis dependencies among checkers
Summary:
This diff detaches analysis dependencies among checkers.

**Problem**: Suppose that we enabled a checker A and B. Then, their abstract semantic functions are registered in a global reference in `callbacks.ml` and used when analyzing a new function/file. However, the problem is that it runs *all of the registered semantic functions* always, even when the analysis is requested only from A. For example, when there is a function call from `foo` to `goo`.

```
void foo() {
  goo();
}
```

If the A checker encounters the function call `goo`, it requests the summary of `goo` which includes the analysis results of both A and B. However, if the B checker has a semantic model of `goo` (or does not need the summary of `goo` for whatever reason), the analysis of B on `goo` is useless, wasting analysis time and memory.

We found this type of problem on some edge cases. For the same target:

* It took X mins when we enabled Pulse.
* It took Y mins when we enabled Starvation.
* However, when enabled both of them it took much bigger time than (X+Y).

**Solution**: To avoid the issue, this diff tries to detach the dependencies during the analysis. This diff introduces the value, `AnalysisRequest.t`, that shows "which checker's summary was exactly requested", so that it can run a subset of registered abstract semantic functions that are *really* necessary/requested.

By this change, a *partial* summary is introduced, which has a subset of payloads/summaries of the registered checkers. This diff is trying to handle them correctly in reading/writing DB and cache.

Note that this diff can save unnecessary analyses on the situation that NOT all of the captured data should be analyzed, e.g. `--changed-files-index` is given.

Reviewed By: ngorogiannis

Differential Revision: D58362879

fbshipit-source-id: 112c6c9f068aad972cc661a6a73fe7c8243f9427